Move routing to app-routing.module
