+++ src/app/components/header/nav/nav.component.html
@@ -1,11 +1,11 @@
 <mat-button-toggle-group name="fontStyle" aria-label="Font Style">
 
-  <mat-button-toggle routerLink="Dashboard">
+  <mat-button-toggle routerLink="Dashboard" routerLinkActive="active">
     <div class="lighter"></div>
     Dashboard
   </mat-button-toggle>
 
-  <mat-button-toggle routerLink="Badges">
+  <mat-button-toggle routerLink="Badges" routerLinkActive="active">
     <div class="lighter"></div>
     Badges
   </mat-button-toggle>
