--- src/app/models/tile-model.ts
@@ -1,6 +1,0 @@
-export interface TileModel {
-    color: string;
-    cols: number;
-    rows: number;
-    text: string;
-}
